Actually finish rename
