+++ NATTunnel-Common/IMessage.cs
@@ -1,0 +1,10 @@
+﻿using System.IO;
+
+namespace NATTunnel.Common
+{
+    public interface IMessage
+    {
+        void Serialize(BinaryWriter writer);
+        void Deserialize(BinaryReader reader);
+    }
+}
