+++ NATTunnel-Common/INodeMessage.cs
