+++ NATTunnel-Common/MessageType.cs
