+++ NATTunnel-Common/MessageTypeAttribute.cs
