+++ NATTunnel-Common/Messages/Data.cs
