+++ NATTunnel-Common/Messages/Disconnect.cs
