+++ NATTunnel-Common/Messages/MasterServerInfoReply.cs
