+++ NATTunnel-Common/Messages/MasterServerInfoRequest.cs
