+++ NATTunnel-Common/Messages/MasterServerPublishRequest.cs
