+++ NATTunnel-Common/Messages/NewConnectionReply.cs
