+++ NATTunnel-Common/Messages/NewConnectionRequest.cs
