+++ NATTunnel-Common/Messages/PingReply.cs
