+++ NATTunnel-Common/Messages/PingRequest.cs
