+++ NATTunnel-Common/Messages/PrintConsole.cs
