+++ NATTunnel-Common/StreamRingBuffer.cs
