+++ NATTunnel.sln
@@ -3,13 +3,13 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.0.32014.148
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel", "DarkTunnel\NATTunnel.csproj", "{37ECCAE3-C70B-4578-9073-00879D5AE539}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel", "NATTunnel\NATTunnel.csproj", "{37ECCAE3-C70B-4578-9073-00879D5AE539}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel-GUI", "DarkTunnel-GUI\NATTunnel-GUI.csproj", "{312CAF19-E129-4AE2-A836-8F6BFF9FBB86}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel-GUI", "NATTunnel-GUI\NATTunnel-GUI.csproj", "{312CAF19-E129-4AE2-A836-8F6BFF9FBB86}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel-Common", "DarkTunnel-Common\NATTunnel-Common.csproj", "{8CF7F29A-4691-4EB5-BBB1-C20FFA06DA7C}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel-Common", "NATTunnel-Common\NATTunnel-Common.csproj", "{8CF7F29A-4691-4EB5-BBB1-C20FFA06DA7C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel-Master", "DarkTunnel-Master\NATTunnel-Master.csproj", "{DEA2FA24-752D-40B0-9621-3B82DE75DC43}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NATTunnel-Master", "NATTunnel-Master\NATTunnel-Master.csproj", "{DEA2FA24-752D-40B0-9621-3B82DE75DC43}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
